pyproject.toml
[tool.poetry]
name = "sql"
version = "0.1.0"
description = ""
authors = ["Bodan1992 <bodan1992@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.0"

[tool.poetry.dev-dependencies]
mysql-connector-python = "^8.0.33"
pyodbc = "^4.0.39"
tabulate = "0.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
